requirements.txt
# ============================================
# CREDIT CARD FRAUD DETECTION - DEPENDENCIES
# ============================================

# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
imbalanced-learn>=0.9.0  # Cho SMOTE và handling imbalanced data

# Kaggle Dataset Loading
kagglehub>=0.2.0  # Load datasets từ Kaggle

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter & Notebooks
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Configuration & Utilities
pyyaml>=5.4.0
joblib>=1.1.0
python-dotenv>=0.19.0

# Model Monitoring & Logging
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Code Quality
black>=21.0
flake8>=4.0.0

# Deep Learning (uncomment nếu cần)
# tensorflow>=2.8.0
# torch>=1.10.0
# keras>=2.8.0

# AutoML (uncomment nếu cần)
# auto-sklearn>=0.14.0
# pycaret>=2.3.0
